Improvements to question index clarity and cleaning up utility functions
Previously there were minor discrepencies between how the question index, names index and round number indexes worked. Now it is clearer... the round index is a value that starts at 1 and goes to the selected round number. The questions and answers arrays are clearly arrays and therefore start at 0 and go to numberOfRounds -1